backend/requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
sqlalchemy
httpx==0.28.1
apscheduler
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.12.5
email-validator==2.3.0
tenacity==9.0.0
